producto mapper modificado

+++ src/main/java/com/GreenCart/GreenCart/persistance/entity/Producto.java
@@ -9,7 +9,7 @@ public class Producto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_producto")
-    private Long idProducto;
+    private Integer  idProducto;
 
     private String nombre;
 
@@ -74,15 +74,14 @@ public class Producto {
         this.nombre = nombre;
     }
 
-    public Long getIdProducto() {
+    public Integer  getIdProducto() {
         return idProducto;
     }
 
-    public void setIdProducto(Long idProducto) {
+    public void setIdProducto(Integer  idProducto) {
         this.idProducto = idProducto;
     }
 
-   
     public Integer getCantidadStock() {
         return cantidadStock;
     }
